chore: wallet daemon and npm pain points (#1288)
Description
---
* In the wallet daemon
* Removed the old Tari connector button (deprecated), leaving only the
WalletConnect option. This is to avoid user confusion.
* Hardcoded the WalletConnect ProjectId as default, to make wallet
daemon binaries automatically use it without needing to specify it in
the envvars (it can still be overridden in the envvars if needed).
* Make `Igor` the default network if no network is specified from CLI
* Adding metadata (repository, homepage and issues) to the npm packages
(`typescript-bindings`, `wallet_jrpc_client`) so they are linked back
from the npm registry. Bumped the version number to trigger a npm
publish.

Motivation and Context
---
We identified some pain points for users/developers related to wallet
daemon and npm packages.

How Has This Been Tested?
---
* Manually by starting a new localnet with tari spawn and connect with
walletconnect
* For the default network value, spawned a wallet daemon directly from
CLI and inspected the logs

What process can a PR reviewer use to test or verify this change?
---
See previous section

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify