fix(tari_validator_node_web_ui): missing templates on template list (#1297)
Description
---
Set `limit` as `0` when fetching templates from the VN JRPC API, which
makes it return all templates

Motivation and Context
---
The list of templates in the VN UI does not show any new templates past
10 (+3 including builtin template).

This PR is a quick fix to ignore the limit and show all templates. In
the future we would want proper pagination by updating the VN JRPC API
to accept an `offset` parameter

How Has This Been Tested?
---
Locally:
* Starting a new localnet with tari_spawn
* Deployed more than 10 different templates
* Checking that all appear on the listing

What process can a PR reviewer use to test or verify this change?
---
See previous section

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify