fix(indexer): new graphql url config option (#1303)
Description
---
* New config option in the indexer and tari_swarm to specify the public
address of the GraphQL API
* The indexer web ui now calls a new` /graphql_address` endpoint to get
the actual GraphQL URL, in similar fashion as the JSON RPC URL.

Motivation and Context
---
When deploying on ContractNet, the events page is empty. This is due to
the GraphQL URL being hardcoded to localhost.

This PR fixes it by parameterizing the public address in a similar way
to what we already do for the JSON RPC public address

How Has This Been Tested?
---
Locally spawning `tari_swarm` and using the indexer web ui

What process can a PR reviewer use to test or verify this change?
---
See previous section

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Introduced a new optional field for the public GraphQL API endpoint
alongside the existing JSON RPC.
- Enhanced the HTTP interface to dynamically provide the GraphQL API
address for improved connectivity.
- Expanded functionality to retrieve and return GraphQL URLs in various
components.

- **Documentation**
- Updated documentation to clarify the usage of JSON RPC and GraphQL
endpoints.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Stan Bondi <sdbondi@users.noreply.github.com>