feat(wallet-daemon): track authored templates and show on UI (#1333)
Description
---
When a transaction is finalized and contains a template that belongs to
any of our account, I add it to a new SQLite table in wallet daemon
mapped by account's key index.
Additionally a new wallet daemon frontend page in order to see what are
the authored templates.


![image](https://github.com/user-attachments/assets/014e9d0d-013a-4bb9-9ef5-5f092176ab96)

![image](https://github.com/user-attachments/assets/653c0df1-a521-4449-b05e-ae8ae8e04fe2)


Motivation and Context
---
It was hard sometimes to see what templates we authored/created with an
account.

How Has This Been Tested?
---
1. Start a fresh swarm
2. Create an account in wallet daemon
3. Publish a template
4. Check new template on templates page

What process can a PR reviewer use to test or verify this change?
---


Breaking Changes
---

- [ ] None
- [x] Requires data directory to be deleted
- [ ] Other - Please specify

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Introduced a dedicated "Templates" section in the wallet application,
allowing users to view their authored templates.
- Enhanced the API to support secure, paginated template listings and
integrated it with a background service for timely updates.
- Updated the user interface with a new menu item, secured route, and
breadcrumb entry for easy navigation to the Templates section.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Stan Bondi <sdbondi@users.noreply.github.com>