feat(wallet): seed word restoration scanner for accounts (#1357)
Description
---
Introduced a scanner service to scan for all of the accounts we may
could have when we restore our wallet.
This operation is done only once when we restore a wallet from seed
words and at the moment only accounts are restored.

Motivation and Context
---
Even if we restore a wallet using seed words, we can't see all the
accounts that are associated with our wallet.

Closes #1336

How Has This Been Tested?
---
Manually

Steps:
1. Start a fresh, empty swarm
2. Create an account in wallet
3. Create some more accounts (like 5-6)
4. Send some tari across the different accounts with different balances
5. On swarm UI click `Delete data` for both indexer and wallet daemon
(so we wont have anything stored)
6. Start both services on Swarm UI
7. Open up wallet daemon UI and wait for accounts and its balances to
appear

What process can a PR reviewer use to test or verify this change?
---

See test

Breaking Changes
---

- [ ] None
- [x] Requires data directory to be deleted
- [ ] Other - Please specify

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Introduced asynchronous readiness endpoints for wallet and indexer
services, allowing clients to verify component availability.
- Added a background resource scanning service to automatically detect
and synchronize wallet accounts.
	- Enabled wallet restoration via an optional seed words configuration.
- Introduced a new method to reset the key manager's index for a
specified branch.

- **Refactor**
	- Enhanced wallet initialization to include a synchronization flag.
- Improved key and template management by making key indices optional
and streamlining data handling.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Stan Bondi <sdbondi@users.noreply.github.com>