feat(testing): add option to compile templates in cargo workspace (#1404)
Description
---
Added an optional custom target directory, so compilation will work even
with a cargo workspace project.

Motivation and Context
---
Compiling templates were only possible in a non-workspace cargo project,
so if a template project is compiled in a different `target` directory
(in workspace case in the root) then the compilation fails because it
tried to load the wasm binary from the wrong place.

How Has This Been Tested?
---
Manually.

What process can a PR reviewer use to test or verify this change?
---


Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Added the ability to specify a custom target directory when compiling
templates.
- Introduced a new way to create template tests with a shared target
directory.

- **Refactor**
- Updated method signatures to support optional target directory
parameters for template addition and testing.
- Streamlined internal logic for template compilation and test
initialization.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->