feat(storage/rocksdb)!: implemented versioned types (#1487)
Description
---
Implements versioned types on the storage layer for these types:
- substates
- transactions
- transaction executions
- proposal and timeout certificates
- epoch checkpoints

Adds a new unit test

Motivation and Context
---
This allows backward-compatible changes to data structures, which are
"lazily" upgraded as they are loaded.
By lazy, we mean that if there is a schema update, we do not iterate
through every record on startup / implement global migrations. Since
there could be millions of entries, we instead encode the version at
which the object was saved, and if/when it is loaded later, we convert
it _infallibly_ to the latest schema.

A unit test was implemented that shows that bincode decodes a type that
has had new variants added after encoding.

How Has This Been Tested?
---
New unit test, manually

What process can a PR reviewer use to test or verify this change?
---
N/A

Breaking Changes
---

- [ ] None
- [x] Requires data directory to be deleted
- [ ] Other - Please specify