docs: rust docs on ResourceManager and FungibleResourceBuilder (#1512)
Description
---

Have fleshed out comments in the Resource Manager and Fungible Resource
Builder as a first draft.

Motivation and Context
---

An objective is to have the code documented in such a fashion as to
automate the generation of user documentation and aid AI agents with
interpretation.

How Has This Been Tested?
---

What process can a PR reviewer use to test or verify this change?
---
Review the provided documentation for accuracy. In particular, it was
heavily implied that the owner of a resource is able to update the
access rules, regardless of the access rules set up, but the default
ResourceAccessRules on FungibleResourceBuilder seems to set the
update_access_rules to DenyAll

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify

---------

Co-authored-by: Stan Bondi <sdbondi@users.noreply.github.com>