feat: add utxo substate (#1517)
Description
---
Adds a UTXO substate
Update ts_rs to 0.11 
Update bindings
Wallet sends less requests to the indexer when submitting transactions

Motivation and Context
---
Adds a UTXO substate type. This substate is unused and will change.

Closes #1146 

There do not appear to be any breaking changes on the db level, since
the utxo substate was added as the last variant of the ID and value
types.

How Has This Been Tested?
---
Checked that things work as before

What process can a PR reviewer use to test or verify this change?
---


Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify