feat: expose new methods for wallet connect link (#1521)
Description
---

Exposes new methods for wallet connect link:
- tari_accountsList
- tari_getAccountByAddress
- tari_getWalletInfo

Motivation and Context
---

Flow editor needs new functionality.
First, it needs to determine the network, that the wallet connection is
bound to. It will be returned as part of `tari_getWalletInfo` response.
Second, the editor has to list available accounts and get their details.
That is what `tari_accountsList` and `tari_getAccountByAddress` are for.

How Has This Been Tested?
---

Not tested. Will test as soon as this is merged and available via
`tari-js` WalletConnect signer.

What process can a PR reviewer use to test or verify this change?
---


Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify