fix: revert "use npm instead of pnpm (#1553)" (#1558)
Description
---
This reverts commit #1553 
Update the README to include installing pnpm

Motivation and Context
---
We use pnpm workspace references to ensure that packages are in sync
with the latest bindings and client packages.

How Has This Been Tested?
---
Revert

What process can a PR reviewer use to test or verify this change?
---


Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify