feat(walletui): improve nft display (#1577)
Description
---
Improves the way the nfts are fetched and displayed in the wallet - it
didn't always reliably fetch the nft list after sending or claiming.
Added a new placeholder, for cases where it's still fetching the nfts.
Fixed pagination after sending, so it doesn't show a blank page if you
sent the last nft on the page.
Added the new address format to the send flow and moved the fee
calculation from the form step, so it'll only happen when you click
"Continue" now.

Motivation and Context
---
NFT UX improvements

How Has This Been Tested?
---
Manually

What process can a PR reviewer use to test or verify this change?
---
Send and claim NFTs - the onscreen updates should be more reliable now.

Breaking Changes
---
x

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify