fix: add an error case for dropped subscriber (#1669)
The previous implementation, if all senders had dropped (should never
happen), the function would create a subscription that would never emit
new items. Although this is completely fine in this case, this may be
surprising in future and potentially lead to bugs.

This PR instead chooses to "crash" the node on initialisation. The error
message indicates to developers that there is a bug.

---------

Co-authored-by: Stan Bondi <sdbondi@users.noreply.github.com>